Move each page's code into its own folder
