+++ client/camera/CameraMain.css
