+++ client/camera/entry.js
