+++ client/camera/helloWorld.js
