+++ client/camera/printPdf.js
