+++ client/editor/entry.js
