+++ client/projector/entry.js
